Phase 4 - Extract petri-model module
Move graph model, PNML and config layers out of monolith into petri-model (depends on petri-math):
- graphnet.*: GraphElement, GraphPetriNet, GraphPetriPlace, GraphPetriTransition,
  GraphArcIn, GraphArcOut, GraphElementIdGenerator
- graphnet.GraphPlace/GraphTransition/GraphArc: moved from graphpresentation
  (pure AWT, no Swing) so graphnet has zero dependency on graphpresentation
- pnml.*: PnmlParser, PnmlGenerator, PnmlConstants, XmlHelper, CoordinateNormalizer,
  ImportResult (PnmlUtils stays in monolith - uses JFileChooser/MessageHelper)
- config.*, utils.NetworkPositionCalculator
- PNML test suite (21 tests) moved to petri-model/src/test

graphpresentation.GraphPlace/GraphTransition → deprecated wrappers extending graphnet counterparts
graphpresentation.GraphArc removed; all usages updated to graphnet.GraphArc
FormulaBuilderServiceImpl: GraphTransition ref updated to GraphPetriTransition

All 21 tests pass.
