+++ petri-app-monolith/pom.xml
@@ -23,6 +23,10 @@
             <artifactId>petri-api</artifactId>
         </dependency>
         <dependency>
+            <groupId>ua.stetsenkoinna</groupId>
+            <artifactId>petri-model</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.netbeans.external</groupId>
             <artifactId>AbsoluteLayout</artifactId>
         </dependency>
