+++ petri-model/src/main/java/ua/stetsenkoinna/config/AppDirectoryType.java
