+++ petri-model/src/main/java/ua/stetsenkoinna/config/FilePathConfig.java
