+++ petri-model/src/main/java/ua/stetsenkoinna/config/ResourcePathConfig.java
