+++ petri-model/src/main/java/ua/stetsenkoinna/graphnet/GraphArcIn.java
@@ -1,7 +1,6 @@
 package ua.stetsenkoinna.graphnet;
 
 import ua.stetsenkoinna.PetriObj.ArcIn;
-import ua.stetsenkoinna.graphpresentation.GraphArc;
 
 import java.awt.BasicStroke;
 import java.awt.Color;
