+++ petri-model/src/main/java/ua/stetsenkoinna/graphnet/GraphElement.java
