+++ petri-model/src/main/java/ua/stetsenkoinna/graphnet/GraphElementIdGenerator.java
