+++ petri-model/src/main/java/ua/stetsenkoinna/graphnet/GraphPetriNet.java
