+++ petri-model/src/main/java/ua/stetsenkoinna/graphnet/GraphPetriPlace.java
@@ -1,7 +1,6 @@
 package ua.stetsenkoinna.graphnet;
 
 import ua.stetsenkoinna.PetriObj.PetriP;
-import ua.stetsenkoinna.graphpresentation.GraphPlace;
 
 import java.awt.Graphics2D;
 import java.io.Serializable;
