+++ petri-model/src/main/java/ua/stetsenkoinna/graphnet/GraphPlace.java
@@ -1,0 +1,61 @@
+package ua.stetsenkoinna.graphnet;
+
+import java.awt.BasicStroke;
+import java.awt.Color;
+import java.awt.Graphics2D;
+import java.awt.geom.Ellipse2D;
+import java.awt.geom.Point2D;
+
+public class GraphPlace extends GraphElement {
+
+    private static int diameter = 40;
+
+    public static int getDiameter() { return diameter; }
+    public static void setDiameter(int aDiameter) { diameter = aDiameter; }
+
+    private Ellipse2D graphElement = new Ellipse2D.Double(0, 0, getDiameter(), getDiameter());
+
+    public GraphPlace() {
+        super.setLineWidth(3);
+        super.setColor(Color.BLACK);
+    }
+
+    @Override
+    public void drawGraphElement(Graphics2D g2) {
+        g2.setStroke(new BasicStroke(getLineWidth()));
+        g2.setColor(getColor());
+        g2.draw(graphElement);
+        g2.setColor(Color.WHITE);
+        g2.fill(graphElement);
+        g2.setColor(getColor());
+    }
+
+    @Override
+    public void setNewCoordinates(Point2D p) {
+        graphElement.setFrame(
+                p.getX() - (double) getDiameter() / 2,
+                p.getY() - (double) getDiameter() / 2,
+                getDiameter(), getDiameter()
+        );
+    }
+
+    @Override
+    public boolean isGraphElement(Point2D p) { return graphElement.contains(p); }
+
+    @Override
+    public Point2D getGraphElementCenter() {
+        return new Point2D.Double(
+                graphElement.getX() + (double) getDiameter() / 2,
+                graphElement.getY() + (double) getDiameter() / 2
+        );
+    }
+
+    @Override
+    public String getType() { return graphElement.getClass().toString(); }
+
+    @Override
+    public int getBorder() { return getDiameter() / 2; }
+
+    public Ellipse2D getGraphElement() { return graphElement; }
+    public void setGraphElement(Ellipse2D graphElement) { this.graphElement = graphElement; }
+}
