+++ petri-model/src/main/java/ua/stetsenkoinna/pnml/CoordinateNormalizer.java
