+++ petri-model/src/main/java/ua/stetsenkoinna/pnml/ImportResult.java
