+++ petri-model/src/main/java/ua/stetsenkoinna/pnml/PnmlConstants.java
