+++ petri-model/src/main/java/ua/stetsenkoinna/pnml/PnmlGenerator.java
