+++ petri-model/src/main/java/ua/stetsenkoinna/pnml/XmlHelper.java
