+++ petri-model/src/main/java/ua/stetsenkoinna/utils/NetworkPositionCalculator.java
