+++ petri-model/src/test/java/pnml/InformationLinkParameterTest.java
