+++ petri-model/src/test/resources/pnml/test3.pnml
