+++ petri-model/src/test/resources/pnml/test_petri_net.pnml
