+++ pom.xml
@@ -13,6 +13,7 @@
         <module>petri-annotation</module>
         <module>petri-math</module>
         <module>petri-api</module>
+        <module>petri-model</module>
         <module>petri-app-monolith</module>
     </modules>
 
@@ -93,6 +94,11 @@
             <dependency>
                 <groupId>ua.stetsenkoinna</groupId>
                 <artifactId>petri-api</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>ua.stetsenkoinna</groupId>
+                <artifactId>petri-model</artifactId>
                 <version>${project.version}</version>
             </dependency>
         </dependencies>
